reco_analysis/pyproject.toml
[project]
name = "reco-analysis"
version = "0.1.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License"
]
authors = [
  { name = "Mike Khor, Annelise Friar, Farid Gholitabar, Gary Kong" },
]
requires-python = "~=3.11"
readme = "README.md"
license = { file = "LICENSE" }

[tool.poetry]
name = "reco-analysis"
version = "0.1.0"
description = "RECO: Recovery Companion, to monitor patients with heart failure on their recovery journey"
authors = [
  "Mike Khor <khorenghoe@gmail.com>",
  "Annelise Friar <anniefriar@berkeley.edu>",
  "Farid Gholitabar <farid.gholitabar@berkeley.edu>",
  "Gary Kong <garykong@berkeley.edu>"
]


[tool.poetry.dependencies]
python = "3.11.7"
awscli = "^1.33.0"
ipython = "^8.25.0"
jupyterlab = "^4.2.1"
loguru = "^0.7.2"
matplotlib = "^3.9.0"
mkdocs = "^1.6.0"
notebook = "^7.2.0"
numpy = "^1.26.4"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.0"
tqdm = "^4.66.4"
typer = "^0.12.3"
transformers = "^4.41.2"
huggingface-hub = "^0.23.3"
ipywidgets = "^8.1.3"
langchain-openai = "^0.1.8"
langchain = "^0.2.3"
langchain-community = "^0.2.4"
streamlit = "^1.35.0"
streamlit-authenticator = "^0.3.2"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
reportlab = "^4.2.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
alembic = "^1.13.1"
pytest = "^8.2.2"
pytest-timeout = "^2.3.1"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff.lint.isort]
known_first_party = ["reco_analysis"]
force_sort_within_sections = true